Fix answer generation for nested dicts
